add: Generate image code file and result image folder
The code file(get no mask image.ipynb) crawls the photo from
here(thispersondoesnotexist.com/image) and screenshots it. The
folder(noMask) is stores the result of the code.

In this project, at the beginning, we want approximately 100 photos of
people with and without mask, this commit does the work that generates
photos without mask.